1) updated the image data info decleration from main to image_data.h
