+++ sw/sdk/ph_2_new/src/imageData.h
@@ -1,0 +1,15 @@
+/*
+ * imageData.h
+ *
+ *  Created on: 8 Jan 2021
+ *      Author: oliver
+ */
+
+#ifndef SRC_IMAGEDATA_H_
+#define SRC_IMAGEDATA_H_
+
+extern unsigned char Test_Image_map[FrameSize];
+extern unsigned char volume_screen[FrameSize];
+extern unsigned char song_selection[FrameSize];
+
+#endif /* SRC_IMAGEDATA_H_ */
