+++ sw/sdk/ph_2_new/src/main.c
@@ -43,7 +43,8 @@
 #include "Bit_Bash.h"
 //#include "Interrupt.h"
 #include "VDMA_Setup.h"
-
+#include "imageData.h"
+#include "GPIO_LEDs.h"
 //#define DEBUG
 
 /*
@@ -56,7 +57,7 @@
 // base address for the DDR3 im memory map
 #define MEM_BASE_ADDR	XPAR_PS7_DDR_0_S_AXI_BASEADDR	//	0x00100000
 
-#define TX_WAV_BUFFER_BASE		(MEM_BASE_ADDR + 0x01000000)
+#define TX_AUDIO_DATA_BUFFER_BASE		(MEM_BASE_ADDR + 0x01000000)
 
 
 #define IMAGE_BUFFER_SIZE 		(MEM_BASE_ADDR + 0x1A000000) //image address in RAM
@@ -80,8 +81,8 @@
  */
 #define RESET_TIMEOUT_COUNTER	10000
 
-// for loop value for LED test
-#define LED_DELAY     10000000
+//// for loop value for LED test
+//#define LED_DELAY     10000000
 
 
 //size of the audio block data sent to the RAM
@@ -97,8 +98,8 @@
 //#define GPIO_BTN_CHANNEL 0
 //#define GPIO_SW_CHANNEL 1
 
-#define SINGLE_LED_CHANNEL 1
-#define RGB_LED_CHANNEL    2
+//#define SINGLE_LED_CHANNEL 1
+//#define RGB_LED_CHANNEL    2
 
 #define BUTTONS_CHANNEL 1
 #define SWITCH_CHANNEL  2
@@ -152,13 +153,14 @@ void I2SSend();
 void I2C_send(u16 address, u8 datasend);
 void CodecReadByte(u16 Address, u8 *BufferPtr, u8 ByteCount);
 
-void Read_SD_Card(char *filename);
-void Check_SD_Card_Mounted(void);
+//void Read_SD_Card(char *filename, size_t AudioBufferSize);
+/*void Check_SD_Card_Mounted(void);*/
 void Open_File_Directory(void);
+
 void Copy_WAV_File_Names(void);
 int Read_SD_Card_short(char *filename);
 
-void turnOffLEDS();
+//void turnOffLEDS();
 void FIFO_Intr_Handler();
 
 int readFile(u32 AudioDataSizeToRead);
@@ -184,13 +186,12 @@ INTC INTC_Inst;
 static XAxiDma AxiDma; /* Instance of the XAxiDma */
 XAxiDma_Config *Config;
 
-// instance of fat file system
-FATFS FS_instance;
+
 // instance of SD card directory
 DIR directory;
 
 // Instances of GPIOs
-XGpio Gpio_LEDs; /* The Instance of the GPIO Driver */
+//XGpio Gpio_LEDs; /* The Instance of the GPIO Driver */
 XGpio Gpio_SWITCH_BUTTON; /* The Instance of the GPIO Driver */
 
 //VARIABLES
@@ -198,17 +199,18 @@ XGpio Gpio_SWITCH_BUTTON; /* The Instance of the GPIO Driver */
 u8 txData[dataCount], rxData[readData], rxData1[readData];
 
 output_t headphones; // structure for setting volume on L&R channel
-RIFF_Chunk_t RIFF_Info;
-fmt_Chunk_t fmt_Chunk; //sampling rate ... etc.
-Data_Chunk_t Data_Chunk;
-FILINFO file_Info; // file names array
 
+FILINFO file_Info; // file names array
+FIL file; // file structure information
+// instance of fat file system
+FATFS FS_instance;
 
 songList_t **songList_ptr;
 
 
-FIL file; // file structure information
+
 UINT nBytesRead = 0;
+
 
 
 //SWITCH AND BUTTON enum
@@ -220,11 +222,7 @@ XAxiVdma AxiVdma;
 //COLOUR RGB;
 
 int numberOfScreens=10;
-//unsigned char Buffer_Image[FrameSize];
-extern unsigned char Test_Image_map[FrameSize];
-extern unsigned char volume_screen[FrameSize];
-extern unsigned char song_selection[FrameSize];
-//unsigned char Pre_Screen_Image[FrameSize];
+
 
 
 volatile u8 TransmitComplete; /* Flag to check completion of Transmission i2c*/
@@ -257,7 +255,7 @@ int songCounterWAV = 0; // number of wav files stored on the SD card
 // memory that will hold the song audio data
 // pointer to AudioBuffer memory space
 //u8 *RxBufferPtr = (u8 *) RX_BUFFER_BASE;
-u8 *AudioBuffer = (u8 *) TX_WAV_BUFFER_BASE;
+u8 *AudioBuffer = (u8 *) TX_AUDIO_DATA_BUFFER_BASE;
 size_t AudioBufferSize = 0;
 
 
@@ -273,6 +271,7 @@ int Status;
 XGpio Gpio_SPI_TFT;
 
 int main(void) {
+
 
 
 	int status;
@@ -302,7 +301,7 @@ int main(void) {
 		return 0;
 	}
 
-	turnOffLEDS(); // set all LEDs to 0;
+	turnOffLEDS(Gpio_LEDs); // set all LEDs to 0;
 
 
 	XIic_SetSendHandler(&I2C, &I2C, (XIic_Handler) I2CSendHandler);
@@ -358,9 +357,15 @@ int main(void) {
 	//*********************************************************************************************
 	// DMA
 	//for(;;) {
-	//check SD card is mounted
-	Check_SD_Card_Mounted();
 
+
+	//check SD card is mounted
+	Check_SD_Card_Mounted(FS_instance);
+
+	// instance of fat file system
+	//	FATFS FS_instance;
+
+	/*FRESULT result = */f_mount(&FS_instance, Drive0Path, MountSDCard);
 	// open the file directory and check files
 	Open_File_Directory();
 
@@ -402,6 +407,7 @@ int main(void) {
 	sprintf(totalSongsPrint,"%d",songCounterWAV);
 
 	while (1) {
+
 		//	tempSongCNT=(songCounterWAV-(songlistsCounter*MAX_SONGS_ON_LIST))+MAX_SONGS_ON_LIST; // top limit
 		switch (SW_STATE) {
 		case DEFAULT:
@@ -1114,205 +1120,7 @@ void I2SSend() {
 	printf("------------------------------------\n\r");
 }
 
-//#############################################################################
-void Read_SD_Card(char *filename) {
-//	RIFF_Chunk_t RIFF_Info;
-//	fmt_Chunk_t fmt_Chunk;
-//	Data_Chunk_t Data_Chunk;
-	FIL file; // file structure information
-	UINT nBytesRead = 0;
 
-	//for testing
-	printf("Read_SD_Card function entered\n\r");
-
-	//File function return code (FRESULT)
-	//result is the instance of (FRESULT) containing the result of the function f_open
-	//FRESULT f_open (FIL* fp, const TCHAR* path, BYTE mode);/* Open or create a file */
-
-	FRESULT result = f_open(&file, filename, FA_READ);
-	if (result != 0) {
-		printf("File not found result = %u\n\r", result);
-	} else {
-		printf("File: %s Found\r\n", filename);
-	}
-
-	// Read the RIFF header
-	// read the file &file
-	// (void*)&RIFF_Info address of array provided
-	// sizeof(RIFF_Info) size of each element of array total
-	//&nBytesRead number of bytes of the array to read
-	result = f_read(&file, (void*) &RIFF_Info, sizeof(RIFF_Info), &nBytesRead);
-
-	// check if read was successful
-	if (result != 0) {
-		printf("Failed to read file\n\r");
-	} else {
-		printf("file read for RIFF header successful\n\r");
-	}
-
-	//compares the first n bytes of structure element RIFF_Info.riff and character array "RIFF".
-	// result
-	if (memcmp(RIFF_Info.ChunkID, "RIFF", 4) != 0) {
-		printf("Illegal file format. RIFF not found\n\r");
-	} else {
-		printf("RIFF_Chunk_t.ChunkID = %.4s\n\r", RIFF_Info.ChunkID);
-	}
-
-	printf("RIFF_Chunk_t.ChunkSize = %lu\n\r",
-			(unsigned long) RIFF_Info.ChunkSize);
-
-	if (memcmp(RIFF_Info.Format, "WAVE", 4) != 0) {
-		printf("Illegal file format. WAVE not found\n\r");
-	} else {
-		printf("RIFF_Chunk_t.ChunkID = %.4s\n\r", RIFF_Info.Format);
-	}
-
-	//the RIFF Chunk reading is complete
-	printf("the RIFF Chunk reading is completed\n\r\n\r\n\r\n\r\n\r\n\r");
-
-	// Read the fmt_Chunk
-	// read the file &file
-	// (void*)&fmt_Info address of array provided
-	// sizeof(fmt_Info) size of each element of array total
-	//&nBytesRead number of bytes of the array to read
-
-	result = f_read(&file, (void*) &fmt_Chunk, sizeof(fmt_Chunk), &nBytesRead);
-	// check if read was successful
-	if (result != 0) {
-		printf("Failed to read file\n\r");
-	} else {
-		printf("file read for fmt header successful\n\r");
-	}
-
-	//check the fmt section via the structure fmt_Chunk_t;
-	// this section holds all of the sampling informtaion for the music
-
-	//u32 Subchunk1_ID;   // big endian       Should contain the letters "fmt " note the extra empty charachter value
-	//u32 Subchunk1_Size; // little endian    16 for PCM.  This is the size of the  rest of the Subchunk which follows this number.
-	//u16 Audio_Format;   // little endian    PCM = 1 (i.e. Linear quantization) Values other than 1 indicate some form of compression.
-	//u16 Num_Channels;   // little endian    Mono = 1, Stereo = 2, etc.
-	//u32 Sample_Rate;    // little endian    8000, 44100, etc.
-	//u32 ByteRate;       // little endian    == SampleRate * NumChannels * BitsPerSample/8
-	//u16 Block_Align;    // little endian    == NumChannels * BitsPerSample/8
-	// The number of bytes for one sample including
-	// all channels.
-	//u16 Bits_Per_Sample;// little endian    8 bits = 8, 16 bits = 16, etc.
-
-	//fmt_Chunk_t fmt_Chunk;
-	// check for the word data in correct position
-	if (memcmp(fmt_Chunk.Subchunk1_ID, "fmt ", 4) != 0) {
-		printf("'fmt' word not found.\n\r");
-		printf("fmt_Chunk.Subchunk1_ID = %s\r\n", fmt_Chunk.Subchunk1_ID);
-	} else {
-		printf("fmt_Chunk.Subchunk1_ID = %.3s\r\n", fmt_Chunk.Subchunk1_ID);
-	}
-
-	//check the size of the sub chunk
-	printf("Subchunk1_Size:\n\r");
-	printf("Data_Chunk.Subchunk1_Size = %u\r\n",
-			(int) Data_Chunk.Subchunk2_size);
-	// https://en.wikipedia.org/wiki/Pulse-code_modulation
-	printf("expected 16 which means PCM (Pulse Code Modulation)\n\r");
-
-	// check the audio format
-	//Type of format (1 is PCM) - 2 byte integer
-	//Audio_Format
-	printf("fmt_Chunk.Audio_Format = %u\n\r", (int) fmt_Chunk.Audio_Format);
-	printf("expected 1 which means PCM \n\r");
-
-	//Num_Channels
-	// for stereo = 2
-	// for mono = 1
-	printf("fmt_Chunk.Num_Channels = %u\r\n", (int) fmt_Chunk.Num_Channels);
-	printf("expected 2 which means stereo  \n\r");
-
-	//Sample_Rate
-	// Sample rate is the number of samples of audio carried per second, measured in Hz or kHz
-	// (one kHz being 1000 Hz). For example, 44100 samples per second can be expressed as
-	// either 44100 Hz, or 44.1 kHz. ... The sample rate determines the maximum audio frequency
-	//that can be reproduced.
-	printf("fmt_Chunk.Sample_Rate = %u\r\n", (int) fmt_Chunk.Sample_Rate);
-	printf("best value should be 44100 that is maximum\n\r");
-
-	//ByteRate
-	//The audio data rate in bytes/sec.
-	// == SampleRate * NumChannels * BitsPerSample/8
-	printf("fmt_Chunk.ByteRate = %lu\r\n", (unsigned long) fmt_Chunk.ByteRate);
-
-	// Block_Align
-	// Block_Align == NumChannels * BitsPerSample/8
-	// The number of bytes for one sample including all channels.
-	// example : For 16-bit PCM format stereo audio, the block alignment value is 4.
-	printf("fmt_Chunk.Block_Align = %u\r\n", (int) fmt_Chunk.Block_Align);
-	printf("block align value should be 4\n\r");
-
-	//u16 Bits_Per_Sample;
-	// little endian    8 bits = 8, 16 bits = 16, etc.
-	// number of bits per sample usually 16 for .WAV
-	printf("fmt_Chunk.Bits_Per_Sample = %u \n\r",
-			(int) fmt_Chunk.Bits_Per_Sample);
-	printf("number of bits per sample = %u \n\r",
-			(int) fmt_Chunk.Bits_Per_Sample);
-
-	// reading Data Chunk file data
-	// read chunk header
-	//Data_Chunk_t Data_Chunk;
-
-	// Read the Data_Chunk header
-	// read the file &file
-	// (void*)&Data_Chunk address of array provided
-	// size of(Data_Chunk) size of each element of array total
-	//&nBytesRead number of bytes of the array to read
-	result = f_read(&file, (void*) &Data_Chunk, sizeof(Data_Chunk),
-			&nBytesRead);
-	if (result != 0) {
-		printf("Failed to read file\n\r");
-	} else {
-		printf("file read\n\r");
-	}
-
-	// check for the word data in correct position
-	if (memcmp(Data_Chunk.Subchunk2_ID, "data", 4) != 0) {
-		printf("data word not found.\n\r");
-		printf("Data_Chunk_t.Subchunk2_ID = %s\r\n", Data_Chunk.Subchunk2_ID);
-	} else {
-		printf("Data_Chunk_t.Subchunk2_ID = %.4s\n\r", Data_Chunk.Subchunk2_ID);
-	}
-
-	//check the size of the song data
-	printf("size of the song data in bytes:\n\r");
-	printf("Data_Chunk.Subchunk2_size = %lu\r\n",
-			(unsigned long) Data_Chunk.Subchunk2_size);
-
-	// audio data to the heap not the ddr
-
-	// size of audio buffer = size of audio song data
-	//AudioBuffer = malloc(Data_Chunk.Subchunk2_size);
-	// if (!AudioBuffer){
-	//	printf("Could not allocate memory\n\r");
-	//}
-	//else{
-	//   printf("audio buffer size changed to match the audio data size as per header file\n\r");
-	// }
-	AudioBufferSize = Data_Chunk.Subchunk2_size;
-	printf(
-			"audio buffer size variable equal to audio data size as per header file\n\r");
-
-	// read the file
-	result = f_read(&file, (void*) AudioBuffer, AudioBufferSize, &nBytesRead);
-	if (result != 0) {
-		printf("Failed to read file\n\r");
-	} else {
-		printf("read file\n\r");
-	}
-	if (nBytesRead != AudioBufferSize) {
-		printf("Didn't read the complete file");
-	} else {
-		printf("read the complete file");
-
-	}
-
-}
 
 //#############################################################################
 int Read_SD_Card_short(char *filename) {
@@ -1615,15 +1423,7 @@ u8 subchunk_SIZE[nbytes];
 	}
 
 }
-//##################################################################################
-void Check_SD_Card_Mounted(void) {
-	FRESULT result = f_mount(&FS_instance, "0:/", 1);
-	if (result != 0) {
-		print("No SD Card mounted, please insert and restart program.\r\n");
-	} else {
-		print("SD Card is mounted\r\n");
-	}
-}
+
 
 //#############################################################################
 void Open_File_Directory(void) {
@@ -2015,60 +1815,7 @@ end:
 
 }
 
-/*****************************************************************************/
-void rgb_test(int LED_RGB_VAL) {
-	volatile int Delay;
-	/* Set the LED to High */
-	XGpio_DiscreteWrite(&Gpio_LEDs, RGB_LED_CHANNEL, LED_RGB_VAL);
 
-#ifdef TESTING
-	print("led  4 ON \n\r");
-#endif
-
-	/* Wait a small amount of time so the LED is visible */
-	for (Delay = 0; Delay < LED_DELAY; Delay++)
-		;
-
-	/* Clear the LED bit */
-	XGpio_DiscreteClear(&Gpio_LEDs, RGB_LED_CHANNEL, LED_RGB_VAL);
-
-#ifdef TESTING
-	print("led  4 OFF \n\r");
-#endif
-
-	/* Wait a small amount of time so the LED is visible */
-	for (Delay = 0; Delay < LED_DELAY; Delay++)
-		;
-}
-/******************************************************************************/
-void Single_LED_test(int LED_NUM) {
-	volatile int Delay;
-	/* Set the LED to High */
-	XGpio_DiscreteWrite(&Gpio_LEDs, SINGLE_LED_CHANNEL, LED_NUM);
-
-#ifdef TESTING
-	print("led  4 ON \n\r");
-#endif
-
-	/* Wait a small amount of time so the LED is visible */
-	for (Delay = 0; Delay < LED_DELAY; Delay++)
-		;
-
-	/* Clear the LED bit */
-	XGpio_DiscreteClear(&Gpio_LEDs, SINGLE_LED_CHANNEL, LED_NUM);
-
-#ifdef TESTING
-	print("led  4 OFF \n\r");
-#endif
-
-	/* Wait a small amount of time so the LED is visible */
-	for (Delay = 0; Delay < LED_DELAY; Delay++)
-		;
-}
-
-void turnOffLEDS(){
-	XGpio_DiscreteWrite(&Gpio_LEDs, SINGLE_LED_CHANNEL, 0x00);
-}
 
 int readFile(u32 AudioDataSizeToRead){
 	return f_read(&file, (void*) AudioBuffer, AudioDataSizeToRead,&nBytesRead);
